+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - properly focusing search input in task sidebar
 - added direct support for Neuralwatt provider
 - added getProviders to extension context to allow fetching the configured providers
+- added truncateToolResult to extension context to allow extensions to use it to truncate the tool results
 
 ## [0.66.0]